requirements.txt
Flask
flask-wtf
flask-cors
python-docx
openai
python-dotenv
PyPDF2
pandas
openpyxl